requirements.txt
altgraph==0.17
appdirs==1.4.4
astroid==2.4.2
builtwith==1.3.4
certifi==2022.12.7
cffi==1.15.0
chardet==3.0.4
colorama==0.4.3
cryptography==3.3.2
decorator==4.4.2
decouple==0.0.7
distlib==0.3.1
filelock==3.0.12
future==0.18.3
idna==2.10
instaloader==4.4.5
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pefile==2019.4.18
pipenv==2022.1.8
psutil==5.7.0
pycparser==2.20
PyInstaller==3.6
pylint==2.5.3
pynput==1.6.8
pyTelegramBotAPI==3.7.1
python-decouple==3.3
python-telegram-bot==12.8
python-telegram-bot[webhooks]==0.25.0
pywin32-ctypes==0.2.0
requests==2.24.0
scapy==2.4.3
six==1.15.0
telegram-send==0.25
toml==0.10.1
tornado==6.0.4
urllib3==1.26.0
virtualenv==20.0.26
virtualenv-clone==0.5.4
wrapt==1.12.1
